# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight REST relay for managing TR-069 devices via the GenieACS NBI. v2.2.0 adds 25 new operational endpoints across 4 phases (CPE lifecycle, NOC support, customer self-service, GenieACS metadata) on top of the v2.1.0 baseline. 100% main-package coverage, 40/40 endpoints end-to-end verified on real ZTE F670L V9.0.10P1N12A via VPN lab (sessions 5i + 5j 2026-04-15).\nSee CHANGELOG.md [2.2.0] for full release notes and upstream genieacs-stack v1.3.1 blocker for customer factory-reset workflows.",
        "title": "GenieACS Relay API",
        "contact": {
            "name": "Cepat Kilat Teknologi",
            "url": "https://github.com/Cepat-Kilat-Teknologi/genieacs-relay"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/genieacs",
    "paths": {
...
        "/devices": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of devices known to GenieACS. Optional filters: model, online, pppoe_username (substring). Used by the admin UI device discovery flow. Pagination metadata: page, page_size, count (rows in this page), has_more (true when count == page_size).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "List devices (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-indexed)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Rows per page (1-200)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model substring (e.g. F670L)",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only include devices that are currently online",
                        "name": "online",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by PPPoE username substring",
                        "name": "pppoe_username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.DevicesListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/devices/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Looks up a single device by MAC address, serial number, or PPPoE username. Exactly one of the three query params must be provided. Returns 404 if no device matches. Used in the customer onboarding flow when the IP is not yet known.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Search for a device by alternative key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC address (case-insensitive, AA:BB:CC:DD:EE:FF format)",
                        "name": "mac",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Serial number from device info",
                        "name": "serial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "PPPoE username (exact match)",
                        "name": "pppoe_username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.DeviceSearchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.DeviceSearchResponse": {
            "description": "Single-device search result. Returns the matching device or 404.",
            "type": "object",
            "properties": {
                "device": {
                    "$ref": "#/definitions/main.DeviceSummary"
                }
            }
        },
...
        "main.DeviceSummary": {
            "description": "Lightweight per-device summary for the list/search endpoints.",
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "last_inform": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                }
            }
        },
        "main.DevicesListResponse": {
            "description": "Paginated device listing response. Pagination metadata is best-effort: page and page_size echo the request, has_more is true when the current page returned exactly page_size rows (i.e. there might be more on the next page).",
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.DeviceSummary"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "main.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "data": {
                    "description": "Response payload (success) OR error detail (string/field list)"
                },
                "error_code": {
                    "description": "Machine-readable error code (VALIDATION_ERROR, NOT_FOUND, ...)",
                    "type": "string"
                },
                "request_id": {
                    "description": "Correlation ID from X-Request-ID header",
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" for 2xx, HTTP reason phrase for errors",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication (required when MIDDLEWARE_AUTH=true)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}